requirements.txt
Django==2.2.12
django-heroku
dj_database_url==0.5.0
psycopg2-binary==2.8.4
django-debug-toolbar==2.2
